Cast to int for python 3

+++ skimage/filter/rank/core16_cy.pyx
@@ -59,7 +59,7 @@ cdef void _core16(dtype_t kernel(Py_ssize_t*, float, dtype_t,
 
     maxbin_list = [0, 0, 4, 8, 16, 32, 64, 128, 256, 512, 1024, 2048, 4096,
                    8192, 16384, 32768, 65536]
-    midbin_list = [m / 2 for m in maxbin_list]
+    midbin_list = [int(m / 2) for m in maxbin_list]
 
     # set maxbin and midbin
     cdef Py_ssize_t maxbin = maxbin_list[bitdepth]
